requirements.txt
requests>=2.32.0
beautifulsoup4>=4.12.0
python-dateutil>=2.9.0
APScheduler>=3.10.4
Flask>=3.0.0

pytest>=8.0.0
